requirements.txt
ajax==3.0.0
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==3.1
django-appconf==1.0.4
django-widget-tweaks==1.4.8
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
joblib==0.17.0
nltk==3.5
Pillow==7.2.0
python-dateutil==1.5
pytz==2020.1
regex==2020.11.13
requests==2.24.0
simplejson==3.17.2
sqlparse==0.3.1
tqdm==4.53.0
urllib3==1.25.10
whitenoise==5.2.0
